+++ lib/event/channel-event.js
