+++ lib/event/event.js
