+++ lib/event/meta-event.js
