+++ lib/event/sysex-event.js
