+++ lib/file/file.js
@@ -3,9 +3,9 @@
 var chunk = require('./chunk');
 var Track = require('./track');
 
-var MetaEvent = require('../events/meta-event');
-var SysexEvent = require('../events/sysex-event');
-var ChannelEvent = require('../events/channel-event');
+var MetaEvent = require('../event/meta-event');
+var SysexEvent = require('../event/sysex-event');
+var ChannelEvent = require('../event/channel-event');
 
 var error = require('../util/error');
 var buffer = require('../util/buffer');
