+++ lib/file/track.js
@@ -5,10 +5,10 @@ var chunk = require('./chunk');
 var buffer = require('../util/buffer');
 var error = require('../util/error');
 
-var Event = require('../events/event');
-var MetaEvent = require('../events/meta-event');
-var SysexEvent = require('../events/sysex-event');
-var ChannelEvent = require('../events/channel-event');
+var Event = require('../event/event');
+var MetaEvent = require('../event/meta-event');
+var SysexEvent = require('../event/sysex-event');
+var ChannelEvent = require('../event/channel-event');
 
 /**
  * Construct a new Track
